:art: extract the API-request handling to a separate middleware
